Tests: check mdsip idle state before setting up server; avoids timing issue
